fixing type

+++ capi/src/qibolab/qibolab.h
@@ -2,4 +2,4 @@
  * Qibolab C API
  */
 
-extern double *execute_qasm(const char *circuit, const char *platform, int nshots);
+extern int *execute_qasm(const char *circuit, const char *platform, int nshots);
